requirements.txt
tensorflow==2.10.1
numpy==1.24.4
matplotlib
pillow
tqdm
requests
scikit-learn
